create file auth
