+++ server/index.js
@@ -6,14 +6,12 @@ const app = express()
 const session = require('express-session')
 const passport = require('passport')
 const bodyParser = require('body-parser')
-const LocalStrategy = require('passport-local').Strategy
 const mypageRouter = require('../routes/mypage')
-const signinRouter = require('../routes/signin')
 const registerRouter = require('../routes/register')
 const createDishRouter = require('../routes/createDish')
 const signoutRouter = require('../routes/signout')
 const callbackRouter = require('../routes/callback')
-const db = require('../models/index')
+const authRouter = require('../routes/auth')
 
 require('dotenv').config()
 
@@ -27,51 +25,14 @@ app.use(
 app.use(passport.initialize())
 app.use(passport.session())
 
-// passportとsessionの紐づけ
-passport.serializeUser(function(username, done) {
-  done(null, username)
-})
-passport.deserializeUser(function(username, done) {
-  done(null, { name: username })
-})
-
-passport.use(
-  'local',
-  new LocalStrategy(
-    {
-      usernameField: 'username',
-      passwordField: 'password'
-    },
-    function(username, password, done) {
-      process.nextTick(() => {
-        db.user
-          .findOne({
-            where: { username, password }
-          })
-          .then((user, err) => {
-            if (err) {
-              return done(err)
-            }
-            if (!user) {
-              return done(null, false, {
-                message: 'usernameまたはpasswordが間違っています'
-              })
-            }
-            return done(null, user)
-          })
-      })
-    }
-  )
-)
-
 app.use(bodyParser.urlencoded({ extended: false }))
 app.use(bodyParser.json())
 app.use('/routes/mypage', mypageRouter)
-app.use('/routes/signin', signinRouter)
 app.use('/routes/signout', signoutRouter)
 app.use('/routes/register', registerRouter)
 app.use('/routes/createDish', createDishRouter)
 app.use('/routes/callback', callbackRouter)
+app.use('/routes/auth', authRouter)
 
 // Import and Set Nuxt.js options
 const config = require('../nuxt.config.js')
